# services/payment/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreatePaymentDTO:
    properties:
      amount:
        type: number
      application_id:
        type: string
      callback_url:
        description: 'Optional: Callback URL for frontend redirect'
        type: string
      currency:
        description: '"IDR", "USD", etc.'
        type: string
      customer_email:
        type: string
      customer_name:
        description: Customer information
        type: string
      customer_phone:
        type: string
      description:
        type: string
      gateway_name:
        description: '"midtrans", "stripe", etc.'
        type: string
      payment_method:
        description: '"credit_card", "bank_transfer", etc.'
        type: string
      user_id:
        type: string
    required:
    - amount
    - application_id
    - currency
    - customer_email
    - customer_name
    - gateway_name
    - payment_method
    - user_id
    type: object
  dto.PaymentResponseDTO:
    properties:
      amount:
        type: number
      application_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      gateway_name:
        type: string
      gateway_order_id:
        type: string
      id:
        type: string
      paid_at:
        type: string
      payment_method:
        type: string
      payment_type:
        type: string
      redirect_url:
        description: URL to complete payment
        type: string
      status:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  handlers.UploadProofReq:
    properties:
      file_id:
        type: string
      file_url:
        type: string
    type: object
  handlers.VerifyPaymentRequest:
    properties:
      action:
        description: '"approve" or "reject"'
        type: string
      admin_id:
        description: Simulated Admin ID
        type: string
      reason:
        description: Required if rejected
        type: string
    type: object
...
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ini adalah dokumentasi API untuk layanan pembayaran YBB Platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: YBB Payment Service API
  version: "1.0"
paths:
...
  /payments:
    post:
      consumes:
      - application/json
      description: Create a new payment intent and get checkout information
      parameters:
      - description: Payment Request Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create Payment Intent
      tags:
      - Payments
  /payments/{id}:
    get:
      description: Get payment details by its ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get Payment Details
      tags:
      - Payments
  /payments/{id}/cancel:
    post:
      description: Cancel a pending payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Cancel Payment
      tags:
      - Payments
  /payments/{id}/proof:
    post:
      consumes:
      - application/json
      description: Menerima URL Bukti Transfer yang sudah diupload oleh API/File Service
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data File (ID & URL)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UploadProofReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Upload Bukti Transfer (Submit URL)
      tags:
      - Manual Payment
  /payments/{id}/refund:
    post:
      description: Refunds a successful payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentResponseDTO'
      summary: Refund Payment
      tags:
      - Payments
  /payments/{id}/retry:
    post:
      consumes:
      - application/json
      description: Create a new transaction based on failed/expired payment
      parameters:
      - description: Old Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Retry Payment
      tags:
      - Payments
  /payments/{id}/verify:
    post:
      consumes:
      - application/json
      description: Admin menyetujui (Approve) atau menolak (Reject) pembayaran manual
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Verifikasi (Action & Admin ID)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Verifikasi Pembayaran (Admin)
      tags:
      - Manual Payment
  /payments/{id}/verify-status:
    post:
      description: Check payment status from Gateway (now disabled/noop)
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Verify Payment Status
      tags:
      - Payments
  /payments/user/{userId}:
    get:
      description: Get list of payments for a user with pagination
      parameters:
      - description: User UUID
        in: path
        name: userId
        required: true
        type: string
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get User's Payments
      tags:
      - Payments
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"